Moves theme files to cartridges directory
Moves theme-related files into a dedicated 'cartridges' directory
for better organization and separation of concerns.

Updates import paths accordingly.

Removes unused action file.
