--- gl-core/actions/index.tsx
@@ -1,3 +1,0 @@
-import { exampleAction } from './exampleAction';
-
-export { exampleAction };
