+++ gl-core/cartridges/Theme/SVGIcons/ChromeIcon.tsx
