+++ gl-core/cartridges/Theme/SVGIcons/DesktopIcon.tsx
