+++ gl-core/cartridges/Theme/SVGIcons/EdgeIcon.tsx
