+++ gl-core/cartridges/Theme/SVGIcons/FirefoxIcon.tsx
