+++ gl-core/cartridges/Theme/SVGIcons/GatsbyIcon.tsx
