+++ gl-core/cartridges/Theme/SVGIcons/GoldenticketIcon.tsx
