+++ gl-core/cartridges/Theme/SVGIcons/GoldlabelOutlined.tsx
