+++ gl-core/cartridges/Theme/SVGIcons/GraphqlIcon.tsx
