+++ gl-core/cartridges/Theme/SVGIcons/LinuxIcon.tsx
