+++ gl-core/cartridges/Theme/SVGIcons/MacIcon.tsx
