+++ gl-core/cartridges/Theme/SVGIcons/MacromediaIcon.tsx
