+++ gl-core/cartridges/Theme/SVGIcons/OpenAIIcon.tsx
