+++ gl-core/cartridges/Theme/SVGIcons/PingpongballIcon.tsx
