+++ gl-core/cartridges/Theme/SVGIcons/SafariIcon.tsx
