+++ gl-core/cartridges/Theme/SVGIcons/WindowsIcon.tsx
