+++ gl-core/cartridges/Theme/SVGIcons/WordpressIcon.tsx
