+++ gl-core/cartridges/Theme/SVGIcons/XboxIcon.tsx
