+++ gl-core/cartridges/Theme/createMUITheme.tsx
@@ -1,4 +1,4 @@
-import config from '../config.json';
+import config from '../../config.json';
 import { createTheme } from '@mui/material';
 // import { TTheme } from '../';
